+++ server.v
@@ -41,9 +41,9 @@ struct App {
 	timeout    i64 // shutdown timeout
 	started_at u64 // start timestamp
 mut:
-	state     shared State // app shared state
-	log       log.Log      // integrated logging
-	log_level log.Level    // logging level
+	state     shared State   // app shared state
+	log       shared log.Log // integrated logging
+	log_level log.Level      // logging level
 pub mut:
 	// db        sqlite.DB
 	logged_in bool // sample, tell if user is logged in
@@ -66,7 +66,9 @@ fn (mut app App) set_app_config() {
 	} $else {
 		app.log_level = log.Level.info
 	}
-	app.log.set_level(app.log_level)
+	lock app.log {
+		app.log.set_level(app.log_level)
+	}
 	println('Logging level set to $app.log_level')
 
 	// app.log.set_full_logpath(log_file)
@@ -77,8 +79,10 @@ fn (mut app App) set_app_metadata() {
 	// get metadata from application module at build time and set in in application
 	// or terminate execution
 	manifest := vmod.decode(@VMOD_FILE) or {
-		app.log.fatal('unable to read V module file')
-		panic(err)
+		lock app.log {
+			app.log.fatal('unable to read V module file')
+		}
+		panic(err) // already called by logger fatal method, but needed by the compiler at the moment
 	}
 	lock app.state {
 		app.state.metadata = {
@@ -101,6 +105,8 @@ fn (mut app App) set_app_metadata() {
 // set_app_static_mappings set application mappings for static content(assets, etc)
 fn (mut app App) set_app_static_mappings() {
 	// map some static content
+	// but first change to executable folder
+	// os.chdir(os.dir(os.executable())) ?
 	// app.handle_static('.', false) // serve static content from current folder
 	app.serve_static('/favicon.ico', 'public/img/favicon.ico')
 	app.serve_static('/css/style.css', 'public/css/style.css')
@@ -143,6 +149,10 @@ fn new_app() &App {
 	}
 	app.log_info('vweb appl, built with V $v_version') // print V version (set at build time)
 
+	$if debug {
+		app.log_info('V compile time debug flag enabled') // print V compile time debug flag, when enabled
+	}
+
 	return app
 }
 
@@ -151,10 +161,10 @@ pub fn (mut app App) before_request() {
 	// url := app.req.url
 	// app.log_debug('${@FN}: url=$url')
 	// app.log_debug('debug test') // temp
-	$if debug { // temporary workaround to logger not writing output in debug here ...
+	$if debug { // small optimization, but in general needed only in performance critical parts ...
 		rlock app.state {
 			msg := '${@FN}: requested total pages: $app.state.cnt_page, total api: $app.state.cnt_api'
-			println(msg) // temp
+			// println(msg) // no more needed now
 			app.log_debug(msg)
 		}
 	}
@@ -164,12 +174,16 @@ pub fn (mut app App) before_request() {
 
 // log_debug log with verbosity debug, using application logger
 fn (mut app App) log_debug(msg string) {
-	app.log.debug(msg)
+	lock app.log {
+		app.log.debug(msg)
+	}
 }
 
 // log_info log with verbosity info, using application logger
 fn (mut app App) log_info(msg string) {
-	app.log.info(msg)
+	lock app.log {
+		app.log.info(msg)
+	}
 }
 
 // inc_cnt_page increment and return counter value for page calls, from shared state
@@ -212,7 +226,7 @@ fn (mut app App) graceful_exit() {
 	time.sleep_ms(app.timeout)
 	exit(0)
 }
- */
+*/
 
 // to_home redirect to home page
 pub fn (mut app App) to_home() vweb.Result {
@@ -230,6 +244,7 @@ pub fn (mut app App) index() vweb.Result {
 // health sample health check route that exposes a fixed json reply at '/health'
 pub fn (mut app App) health() vweb.Result {
 	app.inc_cnt_api() // sample, increment count number of api requests
+	app.log_info(@FN) // sample, to ensure log calls from routes works
 	return app.json('{"statusCode":200, "status":"ok"}')
 	// same as:
 	// app.json('{"statusCode":200, "status":"ok"}')
@@ -239,6 +254,7 @@ pub fn (mut app App) health() vweb.Result {
 // ready sample readiness route that exposes a fixed json reply at '/ready'
 pub fn (mut app App) ready() vweb.Result {
 	app.inc_cnt_api()
+	app.log_info(@FN)
 	// wait for some seconds here, to simulate a real dependencies check (and a slow reply) ...
 	time.sleep(5 * time.second) // wait for 5 seconds
 	return app.json('{"statusCode":200, "status":"ok", 
@@ -263,6 +279,7 @@ pub fn (mut app App) includes() vweb.Result {
 // show headers in the reply (as text), and set a sample cookie
 pub fn (mut app App) cookie() vweb.Result {
 	app.inc_cnt_api()
+	app.log_info(@FN)
 	app.set_cookie(name: 'cookie', value: 'test')
 	return app.text('Headers: TODO') // $app.headers')
 }
@@ -270,18 +287,21 @@ pub fn (mut app App) cookie() vweb.Result {
 // hello sample route that exposes a text reply at '/hello'
 pub fn (mut app App) hello() vweb.Result {
 	app.inc_cnt_api()
+	app.log_info(@FN)
 	return app.text('Hello world from vweb at $time.now().format_ss()')
 }
 
 // hj sample route that exposes a json reply at '/hj'
 pub fn (mut app App) hj() vweb.Result {
 	app.inc_cnt_api()
+	app.log_info(@FN)
 	return app.json('{"Hello":"World"}')
 }
 
 // time sample route that exposes a json reply at '/time'
 pub fn (mut app App) time() vweb.Result {
 	app.inc_cnt_api()
+	app.log_info(@FN)
 	now := time.now()
 	return app.json('{"timestamp":"$now.unix_time()", "time":"$now"}')
 }
@@ -290,6 +310,7 @@ pub fn (mut app App) time() vweb.Result {
 // expected an HTTP error 404 (not found)
 pub fn (mut app App) not_existent() vweb.Result {
 	app.inc_cnt_api()
+	app.log_info(@FN)
 	return app.json('{"msg":"Should not see this reply"}')
 }
 
@@ -298,6 +319,7 @@ pub fn (mut app App) not_existent() vweb.Result {
 ['/user/:id/info']
 pub fn (mut app App) user_info(user string) vweb.Result {
 	app.inc_cnt_api()
+	app.log_info(@FN)
 	return app.json('{"msg":"Hi, it\'s me (user: $user)"}')
 }
 
@@ -305,6 +327,7 @@ pub fn (mut app App) user_info(user string) vweb.Result {
 // (the given code must be a valid code, in the range 100..599)
 pub fn (mut app App) mystatus() vweb.Result {
 	app.inc_cnt_api()
+	app.log_info(@FN)
 	app.set_status(406, 'My error description') // 406 Not Acceptable, as a sample I change here its description in the reply
 	return app.json('{"msg":"My HTTP status code and message"}')
 }
@@ -313,6 +336,7 @@ pub fn (mut app App) mystatus() vweb.Result {
 ['/info']
 pub fn (mut app App) app_info() vweb.Result {
 	app.inc_cnt_api()
+	app.log_info(@FN)
 	// return app.text(app.metadata.str())
 	mut metadata := ''
 	rlock app.state {
@@ -321,6 +345,7 @@ pub fn (mut app App) app_info() vweb.Result {
 		// app.log_debug('${app.state.metadata['name']}-${app.state.metadata['version']} initialized')
 		metadata = '$app.state.metadata'
 	}
+	app.log_debug(@FN)
 	return app.text(metadata)
 }
 
@@ -328,5 +353,6 @@ pub fn (mut app App) app_info() vweb.Result {
 [post]
 pub fn (mut app App) post_dump() vweb.Result {
 	app.inc_cnt_api()
+	app.log_info(@FN)
 	return app.text('Post body: $app.req.data')
 }